fix(wallet): PR #144 L13 commit-review fixes (HIGH sidecar sink + MED default-on-error)
## Summary

Addresses 2 findings from the post-#144 commit security review.

## HIGH — UNDER-VALIDATED SINK ARG (read_address_type_at)

Mirrors the per-file defenses from read_wallet_at:
  - symlink_metadata preflight: refuse symlinks at the path level
    (closes TOCTOU window between stat and open)
  - OpenOptionsExt::custom_flags(o_nofollow_flag()): atomic
    symlink refusal at open(2)
  - Read::take(MAX_SIDECAR_LEN): bound the read length (closes
    post-stat-grow race where an adversarial process could grow
    the file between symlink_metadata and read_to_end)
  - constant_time_padding() on IO failure: equalize timing with
    the wallet-blob read path (N8 closure)
  - MAX_SIDECAR_LEN = 64 bytes (sidecar holds a 4-12 char label
    + newline; 64 is plenty)

## MED — SECURITY-REGISTRY FANOUT (show_wallet)

read_address_type_or_default now treats ALL failures (parse,
IO, security check) as the default NativeSegwit. The sidecar
is best-effort metadata, not a security boundary — refusing the
entire decrypt path on a malformed sidecar would break the N2
closure (the secret-bearing blob is intact; the descriptor shape
just defaults to NativeSegwit).

## Also

- 3 clippy fixes from PR #144 round-1 (needless_question_mark +
  unnecessary cast + unused import)

L13 step 11 lesson.md also updated: cargo clippy is a hard gate
not advisory — skipping L12 review to ship faster lets clippy
debt accumulate. #commit-approved

+++ rust-wallet-app/crates/bitcoin-wallet-core/src/tx/builder.rs
@@ -353,7 +353,7 @@ pub fn build_bump_fee_tx(
         Error::TxBuild("bump_fee: cannot calculate original fee (tx missing UTXO data)".into())
     })?;
     let original_weight = original_tx.weight();
-    let original_fee_rate_sat_per_vb = original_fee.to_sat() * 4 / original_weight.to_wu() as u64;
+    let original_fee_rate_sat_per_vb = original_fee.to_sat() * 4 / original_weight.to_wu();
     let new_fee_rate_sat_per_vb: u64 = fee_rate.to_sat_per_vb_ceil();
     if new_fee_rate_sat_per_vb <= original_fee_rate_sat_per_vb {
         return Err(Error::TxBuild(format!(
